pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "folium>=0.19.4",
    "mysql-connector-python>=9.2.0",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "streamlit-folium>=0.24.0",
    "streamlit>=1.42.1",
    "twilio>=9.4.6",
    "psycopg2-binary>=2.9.10",
    "openpyxl>=3.1.5",
]
